Early client code
This is built off of project 4. I only deleted the buttons and added text fields. This code has three text fields for each word with a label that is supposed to have the scrambled word. Currently the first text field is available but the other two are disabled. As soon as a player enters the correct word, the second text field will be enabled for use, and so on to the third. 

I hard coded the strings to test functionality, until I receive the shuffled strings.