Rename Main.java to Client Code early
This is the early code for the client. 
It is built off of project four with the buttons taken out. 
I added three text fields, two of which are disabled. Once someone correctly unscrambles a word, the next text field in line gets enabled. 

I hard coded the strings for functionality, until I recieve the shuffled string code