Update README.md
Everything is uploaded in regards to the words. Should work for multiple rounds, no duplication every round unless the dictionary is too small, then duplication is allowed. Duplication is allowed every ten rounds (dictionary size is 30). Scrambled words are stored in three different variables to be used in the server logic code/client. 

- Brian