Current Client code
Everything is changeable but for now, we got a scoreboard and gameplay. As soon as someone reaches three rounds, the game is over. As soon as someone reaches three points(for one round), we start a new round with this winner receiving a point out of a possible three indicating he/she has won the round.